New Android conventions migration
Updated:
 - Theme
 - Use Toolbars instead of ActionBars
 - Use PreferenceFragment in a normal Activity
