+++ app/build.gradle
@@ -2,12 +2,13 @@ apply plugin: 'com.android.application'
 
 android {
     compileSdkVersion 23
-    buildToolsVersion "23.0.0"
+    buildToolsVersion "23.0.2"
 
     defaultConfig {
         applicationId "net.czlee.debatekeeper"
         minSdkVersion 11
         targetSdkVersion 23
+        vectorDrawables.useSupportLibrary = true
     }
 
     buildTypes {
@@ -22,5 +23,7 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:support-v4:23.0.0'
+    compile 'com.android.support:appcompat-v7:23.4.0'
+    compile 'com.android.support:support-v4:23.4.0'
+    compile 'com.android.support:support-vector-drawable:23.4.0'
 }
