+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
+<?xml version="1.0" encoding="utf-8"?><!--
  * Copyright (C) 2012 Chuan-Zheng Lee
  *
  * This file is part of the Debatekeeper app, which is licensed under the
@@ -19,7 +18,7 @@
     package="net.czlee.debatekeeper"
     android:installLocation="auto"
     android:versionCode="19"
-    android:versionName="1.1-beta.1" >
+    android:versionName="1.1-beta.1">
 
     <uses-permission android:name="android.permission.VIBRATE" />
     <uses-permission android:name="android.permission.WAKE_LOCK" />
@@ -29,11 +28,11 @@
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/DefaultTheme" >
+        android:theme="@style/Theme.AppCompat.NoActionBar">
         <activity
             android:name=".DebatingActivity"
             android:icon="@drawable/ic_launcher"
-            android:label="@string/app_name" >
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -42,15 +41,31 @@
         </activity>
         <activity
             android:name=".GlobalSettingsActivity"
-            android:label="@string/activityName_GlobalSettings" />
+            android:label="@string/activityName_GlobalSettings"
+            android:parentActivityName=".DebatingActivity">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".DebatingActivity" />
+        </activity>
         <activity
             android:name=".FormatChooserActivity"
-            android:label="@string/activityName_FormatChooser" />
+            android:label="@string/activityName_FormatChooser"
+            android:parentActivityName=".DebatingActivity">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".DebatingActivity" />
+        </activity>
         <activity
             android:name=".PrepTimeBellsEditActivity"
-            android:label="@string/activityName_PrepTimeBellsEdit" />
+            android:label="@string/activityName_PrepTimeBellsEdit"
+            android:parentActivityName=".GlobalSettingsActivity">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".GlobalSettingsActivity" />
+        </activity>
 
         <service android:name=".DebatingTimerService" />
+
     </application>
 
 </manifest>
