+++ app/src/main/res/layout/activity_debate.xml
@@ -22,6 +22,16 @@
     android:orientation="vertical"
     android:background="@android:color/black" >
 
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/mainScreen_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="?attr/colorPrimary"
+        android:elevation="4dp"
+        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+        android:layout_alignParentTop="true"
+        />
+
     <LinearLayout
         android:id="@+id/mainScreen_controlButtons"
         android:layout_width="fill_parent"
@@ -55,12 +65,12 @@
             android:layout_height="fill_parent"
             android:text="@string/mainScreen_playBell_buttonText" />
     </LinearLayout>
-        
+
     <net.czlee.debatekeeper.EnableableViewPager
         android:id="@+id/mainScreen_debateTimerViewPager"
         android:layout_width="fill_parent"
         android:layout_height="fill_parent"
         android:layout_above="@id/mainScreen_controlButtons"
-        android:layout_alignParentTop="true" />
+        android:layout_below="@id/mainScreen_toolbar" />
 
 </RelativeLayout>
