+++ app/src/main/res/layout/activity_format_chooser.xml
@@ -20,6 +20,15 @@
     android:layout_height="match_parent"
     android:orientation="vertical" >
 
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/formatChooser_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="?attr/colorPrimary"
+        android:elevation="4dp"
+        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+        />
+
     <CheckBox
         android:id="@+id/formatChooser_lookForCustomCheckbox"
         android:layout_width="match_parent"
