+++ app/src/main/res/menu/format_chooser_action_bar.xml
@@ -15,12 +15,14 @@
  * You should have received a copy of the GNU General Public Licence
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 -->
-<menu xmlns:android="http://schemas.android.com/apk/res/android" >
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto" >
 
     <item
         android:id="@+id/formatChooser_actionBar_ok"
         android:icon="@drawable/ic_done_white_24dp"
-        android:showAsAction="always" android:title="@string/formatChooser_actionBar_confirmName">
+        app:showAsAction="always"
+        android:title="@string/formatChooser_actionBar_confirmName">
     </item>
 
 </menu>
