+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
+<?xml version="1.0" encoding="utf-8"?><!--
  * Copyright (C) 2012 Phillip Cao, Chuan-Zheng Lee
  *
  * This file is part of the Debatekeeper app, which is licensed under the
@@ -19,7 +18,7 @@
 
     <string name="app_name" translatable="false">Debatekeeper</string>
     <string name="activityName_Debating_withoutFormat" translatable="false">Debatekeeper</string>
-    <string name="activityName_Debating_withFormat" translatable="false">%1$s - Debatekeeper</string>
+    <string name="activityName_Debating_withFormat" translatable="false">%1$s</string>
     <string name="activityName_GlobalSettings">Settings</string>
     <string name="activityName_FormatChooser">Choose style</string>
     <string name="activityName_PrepTimeBellsEdit">Prep timer bells</string>
@@ -96,7 +95,7 @@
     <string name="schemaTooNewDialog_message">This debate style file uses a schema (%1$s) newer than supported (%2$s) by this app version (%3$s). Check for updates on Google Play?</string>
 
     <!-- POI timer -->
-	<string name="poiTimer_moreInfoUrl" translatable="false">https://github.com/czlee/debatekeeper/wiki/Points-of-information</string>
+    <string name="poiTimer_moreInfoUrl" translatable="false">https://github.com/czlee/debatekeeper/wiki/Points-of-information</string>
 
     <!-- Changelog dialog -->
     <!-- By convention, this should be changed to match the manifest versionCode when the changelog changes. -->
@@ -113,4 +112,6 @@
     <!-- Format chooser activity -->
     <string name="formatChooser_lookForCustomCheckbox_text">Look for custom formats (<a href="https://github.com/czlee/debatekeeper/wiki/Writing-your-own-custom-debate-format-file">Learn more</a>)</string>
     <string name="formatChooser_lookForCustom_errorNoReadPermission">You must allow Debatekeeper to read files on your device for custom formats to be read.</string>
+    <string name="title_activity_global_settings">Settings</string>
+    
 </resources>
