--- app/src/main/res/values/themes.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <style name="DefaultTheme" parent="@android:style/Theme.Holo" />
-</resources>
